WEB-1112: Initial partials refactor work
