+++ web/app/containers/checkout-payment/container.js
@@ -1,0 +1,28 @@
+import React from 'react'
+
+// Partials
+import CheckoutPaymentReduxForm from './partials/checkout-payment-form'
+
+// SDK Components
+import {ProgressSteps, ProgressStepsItem} from 'progressive-web-sdk/dist/components/progress-steps'
+
+const CheckoutPayment = () => {
+    return (
+        <div className="t-checkout-payment u-bg-color-neutral-20">
+            <div className="u-bg-color-neutral-10 u-border-light-bottom">
+                <div className="t-checkout-payment__progress">
+                    <ProgressSteps>
+                        <ProgressStepsItem icon="cart-full" title="Cart" href="/checkout/cart/" />
+                        <ProgressStepsItem icon="shipping" title="Shipping" href="/checkout/shipping/" />
+                        <ProgressStepsItem icon="payment-full" title="Payment" current />
+                        <ProgressStepsItem icon="done" title="Done" />
+                    </ProgressSteps>
+                </div>
+            </div>
+
+            <CheckoutPaymentReduxForm />
+        </div>
+    )
+}
+
+export default CheckoutPayment
